refactor: modify SignUpModal to ProfileSettingModal
